+++ tests/test.c
@@ -617,6 +617,8 @@ struct Test tests[] = {
     TEST_CASE(test_lists_keymember),
     TEST_CASE(test_lists_keyfind),
 
+    TEST_CASE_COND(test_reraise, 0, SKIP_STACKTRACES),
+
     // TEST CRASHES HERE: TEST_CASE(memlimit),
 
     { NULL, 0, false, false }